refactor(tauri): 重构后端命令和服务器配置逻辑，提取 ServerState 及各类命令到独立模块以解耦 lib.rs
