+++ tauri-app/src-tauri/src/commands/mod.rs
@@ -1,1 +1,8 @@
 pub mod about;
+pub mod audio;
+pub mod network;
+pub mod system;
+
+pub use audio::*;
+pub use network::*;
+pub use system::*;
